requirements.txt
absl-py==0.9.0
asgiref==3.2.6
astor==0.8.1
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
cryptography==2.8
cycler==0.10.0
dill==0.3.1.1
dj-database-url==0.5.0
Django==3.0.4
django-crontab==0.7.1
django-extensions==2.2.8
django-rest-knox==4.1.0
djangorestframework==3.11.0
Fiona==1.8.13.post1
gast==0.2.2
geopandas==0.7.0
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
joblib==0.14.1
KDEpy==1.0.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.2.1
munch==2.5.0
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.2.0
pandas==1.0.3
pkg-resources==0.0.0
protobuf==3.11.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.6
pyproj==2.6.0
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
sklearn==0.0
sodapy==2.0.0
sqlparse==0.3.1
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.12.1
